backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
openai==1.10.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
openai-whisper==20231117
